Bug fixes, security and websockets

+++ src/main/java/controller/AccountController.java
@@ -12,6 +12,7 @@ import javax.ejb.Stateless;
 import javax.inject.Inject;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
+import javax.ws.rs.HeaderParam;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
@@ -22,10 +23,9 @@ import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import models.Account;
-import models.Role;
-import models.Tweet;
 import service.AccountService;
 import service.TweetService;
+import utils.JwtUtil;
 
 /**
  *
@@ -45,16 +45,18 @@ public class AccountController {
     @Inject
     private TweetService tweetService;
 
+    private JwtUtil jwtUtil = new JwtUtil();
+
     @GET
     public List<Account> get() {
         return accountService.findAll();
     }
 
-    @GET
-    @Path("/{username}/tweet")
-    public List<Tweet> getTweetsByUser(@PathParam("username") String username) {
-        return tweetService.findByUser(username);
-    }
+//    @GET
+//    @Path("/{username}/tweet")
+//    public List<Tweet> getTweetsByUser(@PathParam("username") String username) {
+//        return tweetService.findByUsername(username);
+//    }
 
     @GET
     @Path("/{id}")
@@ -104,12 +106,42 @@ public class AccountController {
         Account user = accountService.findByUsername(context.getCallerPrincipal().getName());
         if (user == null) {
             throw new WebApplicationException(Response.Status.NOT_FOUND);
+
         }
+
         user.setBio(bio);
         user.setWebsiteUrl(websiteURL);
         user.setLocation(location);
 
         accountService.update(user);
+    }
+
+    @PUT
+    @Path("/update")
+    public void updateJWT(@HeaderParam("Authorization") String bearer,
+            @QueryParam("location") String location,
+            @QueryParam("website") String website,
+            @QueryParam("bio") String bio,
+            @QueryParam("email") String email,            
+            @QueryParam("picture") String picture,
+            @QueryParam("username") String username) throws Exception {
+        if (jwtUtil.validateJwt(bearer)) {
+            Account user = accountService.findByUsername(username);
+            if (user == null) {
+                throw new WebApplicationException(Response.Status.NOT_FOUND);
+            }
+            if (!email.isEmpty()) {
+                user.setEmail(email);
+            }
+            user.setBio(bio);
+            user.setWebsiteUrl(website);
+            user.setLocation(location);
+            user.setPicture(picture);
+
+            accountService.update(user);
+        } else {
+            System.out.println("No valid jwt token");
+        }
     }
 
     @PUT
@@ -124,19 +156,54 @@ public class AccountController {
 
     @PUT
     @Path("{id}/following/{followingId}")
-    public void updateFollowing(@PathParam("id") long id, @PathParam("followingId") long followingId) throws Exception {
-        accountService.addFollowing(followingId, id);
+    public void updateFollowing(@HeaderParam("Authorization") String bearer, @PathParam("id") long id, @PathParam("followingId") long followingId) throws Exception {
+        System.out.println(bearer);
+        if (jwtUtil.validateJwt(bearer)) {
+            accountService.addFollowing(followingId, id);
+        } else {
+            System.out.println("No valid jwt token");
+        }
     }
 
     @PUT
     @Path("{id}/follower/{followerId}")
-    public void updateFollower(@PathParam("id") long id, @PathParam("followerId") long followerId) throws Exception {
-        accountService.removeFollowing(followerId, id);
+    public void updateFollower(@HeaderParam("Authorization") String bearer, @PathParam("id") long id, @PathParam("followerId") long followerId) throws Exception {
+        if (jwtUtil.validateJwt(bearer)) {
+            accountService.removeFollowing(followerId, id);
+        } else {
+            System.out.println("No valid jwt token");
+        }
     }
 
     @DELETE
     @Path("/{id}")
     public void delete(@PathParam("id") long id) throws Exception {
-        accountService.delete(id);
+        accountService.deleteById(id);
+    }
+
+    @GET
+    @Path("/following")
+    public List<Account> findFollowing(@QueryParam("username") String username) {
+        Account user = accountService.findByUsername(username);
+        if (user == null) {
+            throw new WebApplicationException(Response.Status.NOT_FOUND);
+        }
+        return accountService.getFollowing(user.getId());
+    }
+
+    @GET
+    @Path("/followers")
+    public List<Account> findFollowers(@QueryParam("username") String username) {
+        Account user = accountService.findByUsername(username);
+        if (user == null) {
+            throw new WebApplicationException(Response.Status.NOT_FOUND);
+        }
+        return accountService.getFollowers(user.getId());
+    }
+
+    @GET
+    @Path("/search")
+    public List<Account> search(@QueryParam("username") String username) {
+        return accountService.search(username);
     }
 }
